fix application startup order
Summary:
using usort to generate a list with unique entries, has the side
effect of sorting that list. This makes it impossible to
manually enforce a start order for optional components.
Replace it with a slighly slower version, that does preserve the
release files ordering.

Test Plan: add a unrelated component to an release and monitor application startup order

Reviewers: orixmann, vtarasen, druss

Reviewed By: druss

Subscribers: aschultz, druss, mkraemer

Maniphest Tasks: T515, T484

Differential Revision: https://review.tpip.net/D1574